Edit student is working as expected
